Changes for WorldTypeTemplate.
- Expose and unify the data properties between WorldType and
  WorldTypeTemplate

- Remove the type getter from template. This prompts a much larger
  discussion about data packs but templates should never reference their
  actual built type. It is an implementation detail and the API should
  declare these as temporary, lifecycle-like objects.
